ajusting retrieve

+++ src/logger.ts
@@ -5,8 +5,10 @@ import Path from 'path';
 
 import Utils from './utils';
 
-export default function setDefaultLogger(context: string) {
-    if (!global.logger?.default) {
+type LoggerName = 'main' | 'error' | 'info' | 'sfdx';
+
+export function setDefaultLogger(context: string) {
+    if (!global.logger?.main) {
         context = context ?? '';
 
         const log4jsDebugMode = !!Utils.getEnvVariable('LOG4JS_DEBUG_MODE');
@@ -32,37 +34,53 @@ export default function setDefaultLogger(context: string) {
             fileName = `debug`;
         }
 
-        let logPath = Path.join(logsDirPath, `${fileName}_${context}.log`);
+        let logPath = global.loggerPath;
 
-        if (log4jsDebugMode) {
-            logPath = Path.join(logsDirPath, 'debug', `${fileName}_${context}.log`);
+        if (!logPath) {
+            Path.join(logsDirPath, `${fileName}_${context}.log`);
 
-            if (Fs.existsSync(Path.join(logsDirPath, 'debug'))) {
-                Fs.mkdirSync(Path.join(logsDirPath, 'debug'), { recursive: true });
-            }
+            if (log4jsDebugMode) {
+                logPath = Path.join(logsDirPath, 'debug', `${fileName}_${context}.log`);
 
-            if (Fs.existsSync(logPath)) {
-                Fs.unlinkSync(logPath);
+                if (Fs.existsSync(Path.join(logsDirPath, 'debug'))) {
+                    Fs.mkdirSync(Path.join(logsDirPath, 'debug'), { recursive: true });
+                }
+
+                if (Fs.existsSync(logPath)) {
+                    Fs.unlinkSync(logPath);
+                }
             }
         }
 
         log4js.configure({
             appenders: {
-                default: { type: 'file', filename: logPath },
-                info: { type: 'console' },
+                main: { type: 'file', filename: logPath },
+                console: { type: 'console' },
             },
             categories: {
-                default: { appenders: ['default'], level: 'all' },
-                info: { appenders: ['default', 'info'], level: 'all' },
-                sfdx: { appenders: ['default', 'info'], level: 'all' },
+                default: { appenders: ['main'], level: 'ALL' },
+                info: { appenders: ['main', 'console'], level: 'ALL' },
+                sfdx: { appenders: ['main'], level: 'ALL' },
+                error: { appenders: ['main', 'console'], level: 'ERROR' },
             },
         });
 
         global.logger = {
-            default: log4js.getLogger('default'),
+            main: log4js.getLogger('default'),
             info: log4js.getLogger('info'),
+            error: log4js.getLogger('error'),
             sfdx: log4js.getLogger('sfdx'),
-            path: logPath,
         };
+
+        global.loggerPath = logPath;
     }
+}
+
+export function getLogger(loggerName: LoggerName) {
+    return global.logger?.[loggerName];
+}
+
+export function functionError(methodPath: string, error: any): any {
+    global.logger?.error.error(`${methodPath} got an error:`);
+    global.logger?.error.error(error);
 }
